Add condition expression assertion for `ExpectedJoinTable` #7880.   * Add `condition` and `columns` assertions for `ExpectedJoinTable`.   * Adjust the old test case, replace joinSpecification with condition or columns in `join-table`.   * Delete unused expected jaxb classes(such as `ExpectedJoinSpecification.java`, `ExpectedAndPredicate.java`).   * Rename `impl.expr` to `impl.expression`.   * Move `ExpectedWhereClause.java` from `impl.predicate` to `impl.where`.
