+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/jaxb/cases/domain/segment/impl/expression/ExpectedBetweenExpression.java
@@ -15,20 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.expr;
+package org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.expression;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.AbstractExpectedSQLSegment;
 
 import javax.xml.bind.annotation.XmlElement;
-import java.util.LinkedList;
-import java.util.List;
 
 @Setter
 @Getter
-public class ExpectedListExpression extends AbstractExpectedSQLSegment implements ExpectedExpressionSegment {
+public class ExpectedBetweenExpression extends AbstractExpectedSQLSegment implements ExpectedExpressionSegment {
+    @XmlElement(name = "not")
+    private boolean not;
 
-    @XmlElement(name = "items")
-    private List<ExpectedExpression> items = new LinkedList<>();
+    @XmlElement(name = "left")
+    private ExpectedExpression left;
+
+    @XmlElement(name = "between-expr")
+    private ExpectedExpression betweenExpr;
+
+    @XmlElement(name = "and-expr")
+    private ExpectedExpression andExpr;
 }
