+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/main/java/org/apache/shardingsphere/test/sql/parser/parameterized/jaxb/cases/domain/segment/impl/expression/complex/ExpectedComplexExpressionSegment.java
@@ -15,17 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.expr.simple;
+package org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.expression.complex;
 
-import lombok.Getter;
-import lombok.Setter;
+import org.apache.shardingsphere.test.sql.parser.parameterized.jaxb.cases.domain.segment.impl.expression.ExpectedExpressionSegment;
 
-import javax.xml.bind.annotation.XmlAttribute;
-
-@Getter
-@Setter
-public final class ExpectedParameterMarkerExpression extends ExpectedBaseSimpleExpression {
+public interface ExpectedComplexExpressionSegment extends ExpectedExpressionSegment {
     
-    @XmlAttribute
-    private int value;
+    /**
+     * Get text.
+     *
+     * @return text
+     */
+    String getText();
 }
