+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/main/resources/case/dml/select-group-by.xml
@@ -122,26 +122,41 @@
                 <right>
                     <simple-table name="t_order_item" alias="i" start-index="51" stop-index="64"/>
                 </right>
-                <joinSpecification>
-                    <and-predicate>
-                        <predicate start-index="69" stop-index="89">
-                            <column-left-value name="user_id" start-index="69" stop-index="77">
-                                <owner name="o" start-index="69" stop-index="69" />
-                            </column-left-value>
-                            <column-right-value name="user_id" start-index="81" stop-index="89">
-                                <owner name="i" start-index="81" stop-index="81" />
-                            </column-right-value>
-                        </predicate>
-                        <predicate start-index="95" stop-index="117">
-                            <column-left-value name="order_id" start-index="95" stop-index="104">
-                                <owner name="o" start-index="95" stop-index="95" />
-                            </column-left-value>
-                            <column-right-value name="order_id" start-index="108" stop-index="117">
-                                <owner name="i" start-index="108" stop-index="108" />
-                            </column-right-value>
-                        </predicate>
-                    </and-predicate>
-                </joinSpecification>
+                <condition>
+                    <binary-operation-expression start-index="69" stop-index="117">
+                        <left>
+                            <binary-operation-expression start-index="69" stop-index="89">
+                                <left>
+                                    <column name="user_id" start-index="69" stop-index="77">
+                                        <owner name="o" start-index="69" stop-index="69" />
+                                    </column>
+                                </left>
+                                <operator>=</operator>
+                                <right>
+                                    <column name="user_id" start-index="81" stop-index="89">
+                                        <owner name="i" start-index="81" stop-index="81" />
+                                    </column>
+                                </right>
+                            </binary-operation-expression>
+                        </left>
+                        <operator>AND</operator>
+                        <right>
+                            <binary-operation-expression start-index="95" stop-index="117">
+                                <left>
+                                    <column name="order_id" start-index="95" stop-index="104">
+                                        <owner name="o" start-index="95" stop-index="95" />
+                                    </column>
+                                </left>
+                                <operator>=</operator>
+                                <right>
+                                    <column name="order_id" start-index="108" stop-index="117">
+                                        <owner name="i" start-index="108" stop-index="108" />
+                                    </column>
+                                </right>
+                            </binary-operation-expression>
+                        </right>
+                    </binary-operation-expression>
+                </condition>
             </join-table>
         </from>
         <projections start-index="7" stop-index="29">
