+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/main/resources/case/dml/select-join.xml
@@ -26,18 +26,21 @@
                 <right>
                     <simple-table name="t_order_item" alias="i" start-index="37" stop-index="50" />
                 </right>
-                <joinSpecification>
-                    <and-predicate>
-                        <predicate start-index="55" stop-index="77">
-                            <column-left-value name="order_id" start-index="55" stop-index="64">
+                <condition>
+                    <binary-operation-expression start-index="55" stop-index="77">
+                        <left>
+                            <column name="order_id" start-index="55" stop-index="64">
                                 <owner name="o" start-index="55" stop-index="55" />
-                            </column-left-value>
-                            <column-right-value name="order_id" start-index="68" stop-index="77">
+                            </column>
+                        </left>
+                        <operator>=</operator>
+                        <right>
+                            <column name="order_id" start-index="68" stop-index="77">
                                 <owner name="i" start-index="68" stop-index="68" />
-                            </column-right-value>
-                        </predicate>
-                    </and-predicate>
-                </joinSpecification>
+                            </column>
+                        </right>
+                    </binary-operation-expression>
+                </condition>
             </join-table>
         </from>
         <projections start-index="7" stop-index="9">
@@ -72,18 +75,21 @@
                 <right>
                     <simple-table name="t_order_item" start-index="40" stop-index="51" />
                 </right>
-                <joinSpecification>
-                    <and-predicate>
-                        <predicate start-index="56" stop-index="95">
-                            <column-left-value name="order_id" start-index="56" stop-index="71">
+                <condition>
+                    <binary-operation-expression start-index="56" stop-index="95">
+                        <left>
+                            <column name="order_id" start-index="56" stop-index="71">
                                 <owner name="t_order" start-index="56" stop-index="62" />
-                            </column-left-value>
-                            <column-right-value name="order_id" start-index="75" stop-index="95">
+                            </column>
+                        </left>
+                        <operator>=</operator>
+                        <right>
+                            <column name="order_id" start-index="75" stop-index="95">
                                 <owner name="t_order_item" start-index="75" stop-index="86" />
-                            </column-right-value>
-                        </predicate>
-                    </and-predicate>
-                </joinSpecification>
+                            </column>
+                        </right>
+                    </binary-operation-expression>
+                </condition>
             </join-table>
         </from>
         <projections start-index="7" stop-index="20">
@@ -118,6 +124,7 @@
                 <right>
                     <simple-table name="t_order_item" alias="i" start-index="31" stop-index="44" />
                 </right>
+                <columns name="order_id" start-index="52" stop-index="59" />
             </join-table>
         </from>
         <projections start-index="7" stop-index="9">
