+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/main/resources/case/dml/select-order-by.xml
@@ -288,26 +288,41 @@
                 <right>
                     <simple-table name="t_order_item" alias="i" start-index="36" stop-index="49" />
                 </right>
-                <joinSpecification>
-                    <and-predicate>
-                        <predicate start-index="54" stop-index="74">
-                            <column-left-value name="user_id" start-index="54" stop-index="62">
-                                <owner name="o" start-index="54" stop-index="54" />
-                            </column-left-value>
-                            <column-right-value name="user_id" start-index="66" stop-index="74">
-                                <owner name="i" start-index="66" stop-index="66" />
-                            </column-right-value>
-                        </predicate>
-                        <predicate start-index="80" stop-index="102">
-                            <column-left-value name="order_id" start-index="80" stop-index="89">
-                                <owner name="o" start-index="80" stop-index="80" />
-                            </column-left-value>
-                            <column-right-value name="order_id" start-index="93" stop-index="102">
-                                <owner name="i" start-index="93" stop-index="93" />
-                            </column-right-value>
-                        </predicate>
-                    </and-predicate>
-                </joinSpecification>
+                <condition>
+                    <binary-operation-expression start-index="54" stop-index="102">
+                        <left>
+                            <binary-operation-expression start-index="54" stop-index="74">
+                                <left>
+                                    <column name="user_id" start-index="54" stop-index="62">
+                                        <owner name="o" start-index="54" stop-index="54" />
+                                    </column>
+                                </left>
+                                <operator>=</operator>
+                                <right>
+                                    <column name="user_id" start-index="66" stop-index="74">
+                                        <owner name="i" start-index="66" stop-index="66" />
+                                    </column>
+                                </right>
+                            </binary-operation-expression>
+                        </left>
+                        <operator>AND</operator>
+                        <right>
+                            <binary-operation-expression start-index="80" stop-index="102">
+                                <left>
+                                    <column name="order_id" start-index="80" stop-index="89">
+                                        <owner name="o" start-index="80" stop-index="80" />
+                                    </column>
+                                </left>
+                                <operator>=</operator>
+                                <right>
+                                    <column name="order_id" start-index="93" stop-index="102">
+                                        <owner name="i" start-index="93" stop-index="93" />
+                                    </column>
+                                </right>
+                            </binary-operation-expression>
+                        </right>
+                    </binary-operation-expression>
+                </condition>
             </join-table>
         </from>
         <projections start-index="7" stop-index="14">
